Extract format specifier classes to a new file
